+++ packages/assets-controller/src/data-sources/index.ts
@@ -6,13 +6,9 @@ export {
 
 export {
   AccountsApiDataSource,
-  createAccountsApiDataSource,
   type AccountsApiDataSourceOptions,
   type AccountsApiDataSourceState,
-  type AccountsApiDataSourceActions,
-  type AccountsApiDataSourceEvents,
-  type AccountsApiDataSourceMessenger,
-  type AccountsApiDataSourceGetAssetsMiddlewareAction,
+  type AccountsApiDataSourceAllowedActions,
 } from './AccountsApiDataSource';
 
 export {
@@ -20,9 +16,6 @@ export {
   createBackendWebsocketDataSource,
   type BackendWebsocketDataSourceOptions,
   type BackendWebsocketDataSourceState,
-  type BackendWebsocketDataSourceActions,
-  type BackendWebsocketDataSourceEvents,
-  type BackendWebsocketDataSourceMessenger,
   type BackendWebsocketDataSourceAllowedActions,
   type BackendWebsocketDataSourceAllowedEvents,
 } from './BackendWebsocketDataSource';
@@ -30,38 +23,23 @@ export {
 export {
   RpcDataSource,
   createRpcDataSource,
+  type RpcDataSourceConfig,
   type RpcDataSourceOptions,
   type RpcDataSourceState,
-  type RpcDataSourceActions,
-  type RpcDataSourceEvents,
-  type RpcDataSourceMessenger,
   type RpcDataSourceAllowedActions,
   type RpcDataSourceAllowedEvents,
   type ChainStatus,
-  type RpcDataSourceGetAssetsMiddlewareAction,
 } from './RpcDataSource';
 
 export {
   TokenDataSource,
   type TokenDataSourceOptions,
-  type TokenDataSourceMessenger,
   type TokenDataSourceAllowedActions,
-  type TokenDataSourceActions,
-  type TokenDataSourceGetAssetsMiddlewareAction,
 } from './TokenDataSource';
 
 export {
   PriceDataSource,
   type PriceDataSourceOptions,
-  type PriceDataSourceMessenger,
-  type PriceDataSourceAllowedActions,
-  type PriceDataSourceActions,
-  type PriceDataSourceEvents,
-  type PriceDataSourceGetAssetsMiddlewareAction,
-  type PriceDataSourceFetchAction,
-  type PriceDataSourceSubscribeAction,
-  type PriceDataSourceUnsubscribeAction,
-  type PriceDataSourceAssetsUpdatedEvent,
 } from './PriceDataSource';
 
 // Unified Snap Data Source (dynamically discovers keyring snaps via PermissionController)
@@ -77,23 +55,6 @@ export {
   // Types
   type SnapDataSourceState,
   type SnapDataSourceOptions,
-  type SnapDataSourceActions,
-  type SnapDataSourceEvents,
-  type SnapDataSourceMessenger,
-  type SnapDataSourceGetAssetsMiddlewareAction,
+  type SnapDataSourceAllowedActions,
+  type SnapDataSourceAllowedEvents,
 } from './SnapDataSource';
-
-// Initialization helpers
-export {
-  initMessengers,
-  initDataSources,
-  type InitMessengersOptions,
-  type InitDataSourcesOptions,
-  type DataSourceMessengers,
-  type DataSources,
-  type DataSourceActions,
-  type DataSourceEvents,
-  type DataSourceAllowedActions,
-  type DataSourceAllowedEvents,
-  type RootMessenger,
-} from './initDataSources';