+++ packages/assets-controller/src/middlewares/index.ts
@@ -1,7 +1,1 @@
-export {
-  DetectionMiddleware,
-  type DetectionMiddlewareOptions,
-  type DetectionMiddlewareMessenger,
-  type DetectionMiddlewareActions,
-  type DetectionMiddlewareGetAssetsMiddlewareAction,
-} from './DetectionMiddleware';
+export { DetectionMiddleware } from './DetectionMiddleware';